passage de usize a u8

+++ src/resolver/heuristic.rs
@@ -6,22 +6,22 @@ pub enum Heuristic {
     Chebyshev,
 }
 
-pub fn distance(start: usize, end: usize) -> usize {
-    (start as isize - end as isize).abs() as usize
+pub fn distance(start: usize, end: usize) -> u16 {
+    (start as i16 - end as i16).abs() as u16
 }
 
-pub fn manathan(dist_x: usize, dist_y: usize) -> usize {
-    dist_x + dist_y
+pub fn manathan(dist_x: u16, dist_y: u16) -> u16 {
+    dist_x as u16 + dist_y as u16
 }
 
-// pub fn euclidienne(dist_x: usize, dist_y: usize) -> usize {
+// pub fn euclidienne(dist_x: u16, dist_y: u16) -> u16 {
 //     ((dist_x.pow(2) + dist_y.pow(2)) as isize).sqrt()
 // }
 
-pub fn chebyshev(dist_x: usize, dist_y: usize) -> usize {
+pub fn chebyshev(dist_x: u16, dist_y: u16) -> u16 {
     (dist_x + dist_y) - min(dist_x, dist_y)
 }
 
-// pub fn octile(dist_x: usize, dist_y: usize) -> usize {
+// pub fn octile(dist_x: u16, dist_y: u16) -> u16 {
 //     (dist_x + dist_y) + ((2 as f64).sqrt() - 2) * min(dist_x, dist_y)
 // }
